users done from vids

+++ 1_tries/2-try/productFunctionality/server/controllers/products.js
@@ -1,60 +1,70 @@
-import Product from '../models/product.js'
-import mongoose from 'mongoose'
+import Product from "../models/product.js";
+import mongoose from "mongoose";
 
+export const getProducts = async (req, res) => {
+  if (!req.userId) return res.json({ message: "Unauthenticated" });
 
-export const getProducts = async(req,res)=>{
-  try{
+  try {
     const products = await Product.find();
 
     console.log(products);
 
     res.status(200).json(products);
-
-  }catch(err){
-    res.status(404).json({message:err.message})
+  } catch (err) {
+    res.status(404).json({ message: err.message });
   }
-}
+};
 
-export const createProduct = async(req,res)=>{
+export const createProduct = async (req, res) => {
+  if (!req.userId) return res.json({ message: "Unauthenticated" });
+
   const prod = req.body;
   const newProduct = new Product(prod);
 
-  try{
-      await newProduct.save();
-      res.status(201).json(newProduct);
-  }catch(err){
-    res.status(409).json({message:err.message})
+  try {
+    await newProduct.save();
+    res.status(201).json(newProduct);
+  } catch (err) {
+    res.status(409).json({ message: err.message });
   }
-}
+};
 
-export const updateProduct = async(req,res)=>{
+export const updateProduct = async (req, res) => {
   // console.log(req.params)
   // console.log(req.body)
-  const {id:_id} = req.params;
+  if (!req.userId) return res.json({ message: "Unauthenticated" });
+
+  const { id: _id } = req.params;
   const prod = req.body;
   // console.log(prod)
 
-  if(!mongoose.Types.ObjectId.isValid(_id)){
+  if (!mongoose.Types.ObjectId.isValid(_id)) {
     return res.status(404).send("No product with that id");
   }
 
-  try{
-    const updatedProduct = await Product.findByIdAndUpdate(_id,{...prod,_id},{new:true});
+  try {
+    const updatedProduct = await Product.findByIdAndUpdate(
+      _id,
+      { ...prod, _id },
+      { new: true }
+    );
     res.json(updatedProduct);
-  }catch(err){
-    console.log(err)
+  } catch (err) {
+    console.log(err);
   }
-}
+};
 
-export const deleteProduct = async (req,res)=>{
-  const {id:_id} = req.params;
-  if(!mongoose.Types.ObjectId.isValid(_id)){
+export const deleteProduct = async (req, res) => {
+  if (!req.userId) return res.json({ message: "Unauthenticated" });
+
+  const { id: _id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(_id)) {
     return res.status(404).send("No product with that id");
   }
 
   await Product.findByIdAndRemove(_id);
 
-  res.json({message:"Product deleted successfully"})
+  res.json({ message: "Product deleted successfully" });
 
   // console.log(_id);
-}
+};
